+++ tech1-framework-b2b-mongodb-security-jwt/src/main/java/io/tech1/framework/b2b/mongodb/security/jwt/resources/BaseSuperAdminResource.java
@@ -3,7 +3,7 @@ package io.tech1.framework.b2b.mongodb.security.jwt.resources;
 import io.tech1.framework.b2b.mongodb.security.jwt.annotations.AbstractFrameworkBaseSecurityResource;
 import io.tech1.framework.b2b.mongodb.security.jwt.domain.dto.responses.ResponseInvitationCode1;
 import io.tech1.framework.b2b.mongodb.security.jwt.domain.dto.responses.ResponseServerSessionsTable;
-import io.tech1.framework.b2b.mongodb.security.jwt.services.SuperAdminService;
+import io.tech1.framework.b2b.mongodb.security.jwt.services.BaseSuperAdminService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,25 +13,24 @@ import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 
-// todo [yy] naming -> BasedSuperAdminResource
 @Slf4j
 @AbstractFrameworkBaseSecurityResource
 @RestController
 @RequestMapping("/superadmin")
 @RequiredArgsConstructor(onConstructor = @__(@Autowired))
-public class SuperAdminResource {
+public class BaseSuperAdminResource {
 
     // Services
-    private final SuperAdminService superAdminService;
+    private final BaseSuperAdminService baseSuperAdminService;
 
     @GetMapping("/invitationCodes/unused")
     public List<ResponseInvitationCode1> getUnusedInvitationCodes() {
-        return this.superAdminService.findUnused();
+        return this.baseSuperAdminService.findUnused();
     }
 
     @GetMapping("/sessions/server")
     public ResponseServerSessionsTable getServerSessions() {
-        return this.superAdminService.getServerSessions();
+        return this.baseSuperAdminService.getServerSessions();
     }
 }