+++ tech1-framework-b2b-mongodb-security-jwt/src/test/java/io/tech1/framework/b2b/mongodb/security/jwt/tests/contexts/TestsApplicationResourcesContext.java
@@ -70,8 +70,8 @@ public class TestsApplicationResourcesContext {
     }
 
     @Bean
-    public SuperAdminService superAdminService() {
-        return mock(SuperAdminService.class);
+    public BaseSuperAdminService superAdminService() {
+        return mock(BaseSuperAdminService.class);
     }
 
     @Bean